Fix 31 validation issues across Phases 115C-G
QA review found 9 HIGH, 20 MEDIUM, and 26 LOW issues. Fixed all
HIGH and MEDIUM items across SQL migrations, C# API, and React UI:

- 115C: rename buying-patterns route, fix top-NAICS ordering, rename
  misleading metric
- 115D: fix SBA type codes (view returned zero rows), JOIN ref_sba_type
  for human-readable cert names, fix relationship direction values,
  rename agencyCode param, require mentor-protege filter, cap hop-1 query
- 115E: validate bulk status updates against allowlist, guard duplicate
  milestone generation, fix calendar weekday alignment, allow Notes clear
- 115F: fix DataGrid duplicate keys, fix revenue formatting, eliminate
  Cartesian join via EXISTS, align profile completeness keys, add PSC
  validation, fix UEI length, invalidate stale queries, handle NO_DATA gap
- 115G: fix DbContext thread-safety (sequential queries), fix MySQL LIMIT
  parameterization, fix div-by-zero in completeness view, remove controller
  double-fetch, add AdminGuard to data quality route, widen RecordsLoaded

+++ api/src/FedProspector.Core/DTOs/Insights/InsightsDtos.cs
@@ -31,7 +31,7 @@ public class DataFreshnessDto
 {
     public string SourceName { get; set; } = string.Empty;
     public DateTime? LastLoadDate { get; set; }
-    public int RecordsLoaded { get; set; }
+    public long RecordsLoaded { get; set; }
     public string? LastLoadStatus { get; set; }
     public int? HoursSinceLastLoad { get; set; }
     public string FreshnessStatus { get; set; } = string.Empty;
